+++ src/isp.c
@@ -39,7 +39,7 @@ void ISP_cmd(Action a)
  * \ingroup     ISP
  * \remarks     
 ******************************************************************************/
-bool ISP_eraseByte(unsigned int addr)
+bool ISP_eraseByte(uint16_t addr)
 {
     /* check address */
     if((addr < ISP_ADDR_START) || (addr > ISP_ADDR_END))
@@ -85,7 +85,7 @@ void ISP_idle(void)
  * \ingroup     ISP
  * \remarks     
 ******************************************************************************/
-byte ISP_readByte(unsigned int addr)
+byte ISP_readByte(uint16_t addr)
 {
     byte dat = 0x00;
 
@@ -110,7 +110,7 @@ byte ISP_readByte(unsigned int addr)
  * \ingroup     ISP
  * \remarks     
 ******************************************************************************/
-void ISP_setAddress(unsigned int addr)
+void ISP_setAddress(uint16_t addr)
 {
     ISP_ADDRL = addr;
     ISP_ADDRH = addr >> 0x8;
@@ -158,7 +158,7 @@ void ISP_trig(void)
  * \ingroup     ISP
  * \remarks     
 ******************************************************************************/
-bool ISP_writeByte(unsigned int addr,byte dat)
+bool ISP_writeByte(uint16_t addr,byte dat)
 {
     if((addr < ISP_ADDR_START) || (addr > ISP_ADDR_END))
     {
